separate Card class and tarot dictionary into distinct modules
