+++ source/backend/CardDictionary.js
@@ -1,0 +1,2203 @@
+var cardDictionary = {
+  0: {
+    name: 'The Fool',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Freeing yourself from limitation',
+      'Expressing joy and youthful vigor',
+      'Being open-minded',
+      'Taking a leap of faith',
+      'Attuning yourself to your instincts',
+      'Being eager or curious',
+      'Exploring your potential',
+      'Embracing innovation and change',
+    ],
+    reversedMeanings: [
+      'Being gullible and naive',
+      'Taking unnecessary risks',
+      'Failing to be serious when required',
+      'Being silly or distracted',
+      'Lacking experience',
+      'Failing to honor well-established traditions and limits',
+      'Behaving inappropriately',
+    ],
+    suit: 'Trump',
+    keywords: ['freedom', 'faith', 'inexperience', 'innocence'],
+    img: 'm00.jpg',
+    symbolism: 'TODO',
+    description:
+      'With light step, as if earth and its trammels had little power to restrain him, a young man in gorgeous vestments pauses at the brink of a precipice among the great heights of the world; he surveys the blue distance before him-its expanse of sky rather than the prospect below. His act of eager walking is still indicated, though he is stationary at the given moment; his dog is still bounding. The edge which opens on the depth has no terror; it is as if angels were waiting to uphold him, if it came about that he leaped from the height. His countenance is full of intelligence and expectant dream. He has a rose in one hand and in the other a costly wand, from which depends over his right shoulder a wallet curiously embroidered. He is a prince of the other world on his travels through this one-all amidst the morning glory, in the keen air. The sun, which shines behind him, knows whence he came, whither he is going, and how he will return by another path after many days. He is the spirit in search of experience. Many symbols of the Instituted Mysteries are summarized in this card, which reverses, under high warrants, all the confusions that have preceded it.In his Manual of Cartomancy, Grand Orient has a curious suggestion of the office of Mystic Fool, as apart of his process in higher divination; but it might call for more than ordinary gifts to put it into operation. We shall see how the card fares according to the common arts of fortune-telling, and it will be an example, to those who can discern, of the fact, otherwise so evident, that the Trumps Major had no place originally in the arts of psychic gambling, when cards are used as the counters and pretexts. Of the circumstances under which this art arose we know, however, very little. The conventional explanations say that the Fool signifies the flesh, the sensitive life, and by a peculiar satire its subsidiary name was at one time the alchemist, as depicting folly at the most insensate stage.',
+    numeral: '0',
+  },
+  1: {
+    name: 'The Magician',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Taking appropriate action',
+      'Receiving guidance from a higher power',
+      'Becoming a channel of divine will',
+      'Expressing masculine energy in appropriate and constructive ways',
+      'Being yourself in every way',
+    ],
+    reversedMeanings: [
+      'Inflating your own ego',
+      'Abusing talents',
+      'Manipulating or deceiving others',
+      'Being too aggressive',
+      'Using cheap illusions to dazzle others',
+      'Refusing to invest the time and effort needed to master your craft',
+      'Taking shortcuts',
+    ],
+    suit: 'Trump',
+    keywords: ['capability', 'empowerment', 'activity'],
+    img: 'm01.jpg',
+    symbolism: 'TODO',
+    description:
+      'A youthful figure in the robe of a magician, having the countenance of divine Apollo, with smile of confidence and shining eyes. Above his head is the mysterious sign of the Holy Spirit, the sign of life, like an endless cord, forming the figure 8 in a horizontal position . About his waist is a serpent-cincture, the serpent appearing to devour its own tail. This is familiar to most as a conventional symbol of eternity, but here it indicates more especially the eternity of attainment in the spirit. In the Magician\'s right hand is a wand raised towards heaven, while the left hand is pointing to the earth. This dual sign is known in very high grades of the Instituted Mysteries; it shews the descent of grace, virtue and light, drawn from things above and derived to things below. The suggestion throughout is therefore the possession and communication of the Powers and Gifts of the Spirit. On the table in front of the Magician are the symbols of the four Tarot suits, signifying the elements of natural life, which lie like counters before the adept, and he adapts them as he wills. Beneath are roses and lilies, the flos campi and lilium convallium, changed into garden flowers, to shew the culture of aspiration. This card signifies the divine motive in man, reflecting God, the will in the liberation of its union with that which is above. It is also the unity of individual being on all planes, and in a very high sense it is thought, in the fixation thereof. With further reference to what I have called the sign of life and its connexion with the number 8, it may be remembered that Christian Gnosticism speaks of rebirth in Christ as a change "unto the Ogdoad." The mystic number is termed Jerusalem above, the Land flowing with Milk and Honey, the Holy Spirit and the Land of the Lord. According to Martinism, 8 is the number of Christ.',
+    numeral: '1',
+  },
+  2: {
+    name: 'The High Priestess',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Listening to your feelings and intuitions',
+      'Exploring unconventional spirituality',
+      'Keeping secrets',
+      'Being receptive',
+      'Reflecting instead of acting',
+      'Observing others',
+      'Preserving purity',
+    ],
+    reversedMeanings: [
+      'Being aloof',
+      'Obsessing on secrets and conspiracies',
+      'Rejecting guidance from spirit or intuition',
+      'Revealing all',
+      'Ignoring gut feelings',
+      'Refusing to become involved, even when involvement is appropriate',
+    ],
+    suit: 'Trump',
+    keywords: ['intuition', 'reflection', 'purity', 'initiation'],
+    img: 'm02.jpg',
+    symbolism: 'TODO',
+    description:
+      'She has the lunar crescent at her feet, a horned diadem on her head, with a globe in the middle place, and a large solar cross on her breast. The scroll in her hands is inscribed with the word Tora, signifying the Greater Law, the Secret Law and the second sense of the Word. It is partly covered by her mantle, to shew that some things are implied and some spoken. She is seated between the white and black pillars--J. and B.--of the mystic Temple, and the veil of the Temple is behind her: it is embroidered with palms and pomegranates. The vestments are flowing and gauzy, and the mantle suggests light--a shimmering radiance. She has been called occult Science on the threshold of the Sanctuary of Isis, but she is really the Secret Church, the House which is of God and man. She represents also the Second Marriage of the Prince who is no longer of this world; she is the spiritual Bride and Mother, the daughter of the stars and the Higher Garden of Eden. She is, in fine, the Queen of the borrowed light, but this is the light of all. She is the Moon nourished by the milk of the Supernal Mother.In a manner, she is also the Supernal Mother herself--that is to say, she is the bright reflection. It is in this sense of reflection that her truest and highest name in bolism is Shekinah--the co-habiting glory. According to Kabalism, there is a Shekinah both above and below. In the superior world it is called Binah, the Supernal Understanding which reflects to the emanations that are beneath. In the lower world it is MaIkuth--that world being, for this purpose, understood as a blessed Kingdom that with which it is made blessed being the Indwelling Glory. Mystically speaking, the Shekinah is the Spiritual Bride of the just man, and when he reads the Law she gives the Divine meaning. There are some respects in which this card is the highest and holiest of the Greater Arcana.',
+    numeral: '2',
+  },
+  3: {
+    name: 'The Empress',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Nurturing yourself and others',
+      'Bearing fruit',
+      'Celebrating your body',
+      'Bearing (literal or figurative) children',
+      'Reveling in luxury',
+      'Mothering those around you in positive ways',
+      'Enjoying your sexuality',
+      'Getting things done',
+    ],
+    reversedMeanings: [
+      'Overindulging',
+      'Being greedy',
+      'Smothering someone with attention',
+      'Debilitating someone by being overprotective',
+      'Inhibiting productivity by obsessing on productivity',
+      'Being overcome by addictive behavior',
+    ],
+    suit: 'Trump',
+    keywords: ['fertility', 'productivity', 'ripeness', 'nurturing'],
+    img: 'm03.jpg',
+    symbolism: 'TODO',
+    description:
+      'A stately figure, seated, having rich vestments and royal aspect, as of a daughter of heaven and earth. Her diadem is of twelve stars, gathered in a cluster. The symbol of Venus is on the shield which rests near her. A field of corn is ripening in front of her, and beyond there is a fall of water. The sceptre which she bears is surmounted by the globe of this world. She is the inferior Garden of Eden, the Earthly Paradise, all that is symbolized by the visible house of man. She is not Regina coeli, but she is still refugium peccatorum, the fruitful mother of thousands. There are also certain aspects in which she has been correctly described as desire and the wings thereof, as the woman clothed with the sun, as Gloria Mundi and the veil of the Sanctum Sanctorum; but she is not, I may add, the soul that has attained wings, unless all the symbolism is counted up another and unusual way. She is above all things universal fecundity and the outer sense of the Word. This is obvious, because there is no direct message which has been given to man like that which is borne by woman; but she does not herself carry its interpretation.In another order of ideas, the card of the Empress signifies the door or gate by which an entrance is obtained into this life, as into the Garden of Venus; and then the way which leads out therefrom, into that which is beyond, is the secret known to the High Priestess: it is communicated by her to the elect. Most old attributions of this card are completely wrong on the symbolism--as, for example, its identification with the Word, Divine Nature, the Triad, and so forth.',
+    numeral: '3',
+  },
+  4: {
+    name: 'The Emperor',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Exercising authority',
+      'Defining limits',
+      'Directing the flow of work',
+      'Communicating clear guidelines',
+      'Being in control of yourself and others',
+      'Tempering aggressive masculinity with wisdom and experience',
+    ],
+    reversedMeanings: [
+      'Micromanaging',
+      'Crushing the creativity of others with a rigid, iron-fisted approach',
+      'Insisting on getting your own way',
+      'Assuming a dictatorial mindset',
+      'Using overt force to achieve your goals and maintain order',
+    ],
+    suit: 'Trump',
+    keywords: ['authority', 'regulation', 'direction', 'structure'],
+    img: 'm04.jpg',
+    symbolism: 'TODO',
+    description:
+      'He has a form of the Crux ansata for his sceptre and a globe in his left hand. He is a crowned monarch--commanding, stately, seated on a throne, the arms of which axe fronted by rams\' heads. He is executive and realization, the power of this world, here clothed with the highest of its natural attributes. He is occasionally represented as seated on a cubic stone, which, however, confuses some of the issues. He is the virile power, to which the Empress responds, and in this sense is he who seeks to remove the Veil of Isis; yet she remains virgo intacta.It should be understood that this card and that of the Empress do not precisely represent the condition of married life, though this state is implied. On the surface, as I have indicated, they stand for mundane royalty, uplifted on the seats of the mighty; but above this there is the suggestion of another presence. They signify also--and the male figure especially--the higher kingship, occupying the intellectual throne. Hereof is the lordship of thought rather than of the animal world. Both personalities, after their own manner, are "full of strange experience," but theirs is not consciously the wisdom which draws from a higher world. The Emperor has been described as (a) will in its embodied form, but this is only one of its applications, and (b) as an expression of virtualities contained in the Absolute Being--but this is fantasy.',
+    numeral: '4',
+  },
+  5: {
+    name: 'The Hierophant',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Teaching or guiding others',
+      'Searching for the truth',
+      'Asking for guidance from a higher power',
+      'Acknowledging the wisdom and experience of others',
+      'Taking vows',
+      'Engaging in heartfelt rituals',
+      'Volunteering',
+    ],
+    reversedMeanings: [
+      'Using experience as a means of manipulating or misguiding others',
+      'Being dogmatic',
+      'Favoring tradition over what is expedient or necessary',
+      'Going through the motions of empty rituals',
+      'Concealing wisdom',
+      'Restricting access to spiritual truths or the gods',
+    ],
+    suit: 'Trump',
+    keywords: ['guidance', 'knowledge', 'revelation', 'belief'],
+    img: 'm05.jpg',
+    symbolism: 'TODO',
+    description:
+      'He wears the triple crown and is seated between two pillars, but they are not those of the Temple which is guarded by the High Priestess. In his left hand he holds a sceptre terminating in the triple cross, and with his right hand he gives the well-known ecclesiastical sign which is called that of esotericism, distinguishing between the manifest and concealed part of doctrine. It is noticeable in this connexion that the High Priestess makes no sign. At his feet are the crossed keys, and two priestly ministers in albs kneel before him. He has been usually called the Pope, which is a particular application of the more general office that he symbolizes. He is the ruling power of external religion, as the High Priestess is the prevailing genius of the esoteric, withdrawn power. The proper meanings of this card have suffered woeful admixture from nearly all hands. Grand Orient says truly that the Hierophant is the power of the keys, exoteric orthodox doctrine, and the outer side of the life which leads to the doctrine; but he is certainly not the prince of occult doctrine, as another commentator has suggested.He is rather the summa totius theologiæ, when it has passed into the utmost rigidity of expression; but he symbolizes also all things that are righteous and sacred on the manifest side. As such, he is the channel of grace belonging to the world of institution as distinct from that of Nature, and he is the leader of salvation for the human race at large. He is the order and the head of the recognized hierarchy, which is the reflection of another and greater hierarchic order; but it may so happen that the pontiff forgets the significance of this his symbolic state and acts as if he contained within his proper measures all that his sign signifies or his symbol seeks to shew forth. He is not, as it has been thought, philosophy-except on the theological side; he is not inspiration; and he is not religion, although he is a mode of its expression.',
+    numeral: '5',
+  },
+  6: {
+    name: 'The Lovers',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Being in love',
+      'Showing your love to others',
+      'Expressing passion or romantic feelings',
+      'Aligning yourself with groups or like-minded others',
+      'Bringing people together',
+      'Making well-informed decisions',
+    ],
+    reversedMeanings: [
+      'Debilitating passion',
+      'Allowing an unhealthy desire for love to motivate destructive behavior',
+      'Disrupting unity',
+      'Working against the best interests of those who care about you',
+      'Ill-informed decisions',
+    ],
+    suit: 'Trump',
+    keywords: ['love', 'passion', 'unity', 'choice'],
+    img: 'm06.jpg',
+    symbolism: 'TODO',
+    description:
+      'The sun shines in the zenith, and beneath is a great winged figure with arms extended, pouring down influences. In the foreground are two human figures, male and female, unveiled before each other, as if Adam and Eve when they first occupied the paradise of the earthly body. Behind the man is the Tree of Life, bearing twelve fruits, and the Tree of the Knowledge of Good and Evil is behind the woman; the serpent is twining round it. The figures suggest youth, virginity, innocence and love before it is contaminated by gross material desire. This is in all simplicity the card of human love, here exhibited as part of the way, the truth and the life. It replaces, by recourse to first principles, the old card of marriage, which I have described previously, and the later follies which depicted man between vice and virtue. In a very high sense, the card is a mystery of the Covenant and Sabbath.The suggestion in respect of the woman is that she signifies that attraction towards the sensitive life which carries within it the idea of the Fall of Man, but she is rather the working of a Secret Law of Providence than a willing and conscious temptress. It is through her imputed lapse that man shall arise ultimately, and only by her can he complete himself. The card is therefore in its way another intimation concerning the great mystery of womanhood. The old meanings fall to pieces of necessity with the old pictures, but even as interpretations of the latter, some of them were of the order of commonplace and others were false in symbolism.',
+    numeral: '6',
+  },
+  7: {
+    name: 'The Chariot',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Breaking through barriers',
+      'Moving forward with confidence and authority',
+      'Reaching the pinnacle of success',
+      'Basking in the glory of achievement',
+      'Guiding an effort to total victory',
+      'Establishing yourself as a worthy leader',
+    ],
+    reversedMeanings: [
+      'Resting on laurels',
+      'Riding roughshod over the feelings or expectations of others',
+      'Focusing more on past successes than future opportunities',
+      'Failing to rein in impulsive behavior',
+    ],
+    suit: 'Trump',
+    keywords: ['advancement', 'victory', 'triumph', 'success'],
+    img: 'm07.jpg',
+    symbolism: 'TODO',
+    description:
+      'An erect and princely figure carrying a drawn sword and corresponding, broadly speaking, to the traditional description which I have given in the first part. On the shoulders of the victorious hero are supposed to be the Urim and Thummim. He has led captivity captive; he is conquest on all planes--in the mind, in science, in progress, in certain trials of initiation. He has thus replied to the sphinx, and it is on this account that I have accepted the variation of Éliphas Lévi; two sphinxes thus draw his chariot. He is above all things triumph in the mind.It is to be understood for this reason (a) that the question of the sphinx is concerned with a Mystery of Nature and not of the world of Grace, to which the charioteer could offer no answer; (b) that the planes of his conquest are manifest or external and not within himself; (c) that the liberation which he effects may leave himself in the bondage of the logical understanding; (d) that the tests of initiation through which he has passed in triumph are to be understood physically or rationally; and (e) that if he came to the pillars of that Temple between which the High Priestess is seated, he could not open the scroll called Tora, nor if she questioned him could he answer. He is not hereditary royalty and he is not priesthood.',
+    numeral: '7',
+  },
+  8: {
+    name: 'Strength',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Imposing restrictions on yourself for your own benefit',
+      'Bringing your passions under the control of reason',
+      'Resisting impulses that work against your best interests',
+      'Taking bold action',
+    ],
+    reversedMeanings: [
+      'Indulging weakness, even when you know it will damage your health and happiness',
+      'Languishing in addiction',
+      'Allowing your instincts to tame and conquer you',
+      'Failing to take a stand when necessary',
+    ],
+    suit: 'Trump',
+    keywords: [
+      'discipline',
+      'boldness',
+      'self-discipline',
+      'power',
+      'vitality',
+    ],
+    img: 'm08.jpg',
+    symbolism: 'TODO',
+    description:
+      'A woman, over whose head there broods the same symbol of life which we have seen in the card of the Magician, is closing the jaws of a lion. The only point in which this design differs from the conventional presentations is that her beneficent fortitude has already subdued the lion, which is being led by a chain of flowers. For reasons which satisfy myself, this card has been interchanged with that of justice, which is usually numbered eight. As the variation carries nothing with it which will signify to the reader, there is no cause for explanation. Fortitude, in one of its most exalted aspects, is connected with the Divine Mystery of Union; the virtue, of course, operates in all planes, and hence draws on all in its symbolism. It connects also with innocentia inviolata, and with the strength which resides in contemplation.These higher meanings are, however, matters of inference, and I do not suggest that they are transparent on the surface of the card. They are intimated in a concealed manner by the chain of flowers, which signifies, among many other things, the sweet yoke and the light burden of Divine Law, when it has been taken into the heart of hearts. The card has nothing to do with self-confidence in the ordinary sense, though this has been suggested--but it concerns the confidence of those whose strength is God, who have found their refuge in Him. There is one aspect in which the lion signifies the passions, and she who is called Strength is the higher nature in its liberation. It has walked upon the asp and the basilisk and has trodden down the lion and the dragon.',
+    numeral: '8',
+  },
+  9: {
+    name: 'The Hermit',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Becoming or seeking out a guru',
+      'Going on a retreat',
+      'Recharging spiritual or creative batteries',
+      'Lighting the way for those with less experience',
+      'Stepping back to gain perspective',
+    ],
+    reversedMeanings: [
+      'Being a loner',
+      'Fearing contact with others',
+      'Becoming a know-it-all',
+      'Inflating claims of expertise',
+      'Hiding your skills and talents out of fear of unworthiness',
+    ],
+    suit: 'Trump',
+    keywords: ['solitude', 'experience', 'stillness', 'withdrawal'],
+    img: 'm09.jpg',
+    symbolism: 'TODO',
+    description:
+      'The variation from the conventional models in this card is only that the lamp is not enveloped partially in the mantle of its bearer, who blends the idea of the Ancient of Days with the Light of the World It is a star which shines in the lantern. I have said that this is a card of attainment, and to extend this conception the figure is seen holding up his beacon on an eminence. Therefore the Hermit is not, as Court de Gebelin explained, a wise man in search of truth and justice; nor is he, as a later explanation proposes, an especial example of experience. His beacon intimates that "where I am, you also may be."It is further a card which is understood quite incorrectly when it is connected with the idea of occult isolation, as the protection of personal magnetism against admixture. This is one of the frivolous renderings which we owe to Éliphas Lévi. It has been adopted by the French Order of Martinism and some of us have heard a great deal of the Silent and Unknown Philosophy enveloped by his mantle from the knowledge of the profane. In true Martinism, the significance of the term Philosophe inconnu was of another order. It did not refer to the intended concealment of the Instituted Mysteries, much less of their substitutes, but--like the card itself--to the truth that the Divine Mysteries secure their own protection from those who are unprepared.',
+    numeral: '9',
+  },
+  10: {
+    name: 'Wheel of Fortune',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Allowing events to unfold',
+      'Seeing the larger pattern in everyday events',
+      'Trusting your luck',
+      'Watching for cycles',
+      'Believing that "what goes around, comes around"',
+    ],
+    reversedMeanings: [
+      'Losing money gambling',
+      'Refusing to do your part to bring a plan to fruition',
+      'Taking a fatalistic approach to life',
+      'Fighting the natural course of events',
+    ],
+    suit: 'Trump',
+    keywords: ['luck', 'randomness', 'cycles', 'karma', 'fate', 'revolution'],
+    img: 'm10.jpg',
+    symbolism: 'TODO',
+    description:
+      "In this symbol I have again followed the reconstruction of \u00c9liphas L\u00e9vi, who has furnished several variants. It is legitimate--as I have intimated--to use Egyptian symbolism when this serves our purpose, provided that no theory of origin is implied therein. I have, however, presented Typhon in his serpent form. The symbolism is, of course, not exclusively Egyptian, as the four Living Creatures of Ezekiel occupy the angles of the card, and the wheel itself follows other indications of L\u00e9vi in respect of Ezekiel's vision, as illustrative of the particular Tarot Key. With the French occultist, and in the design itself, the symbolic picture stands for the perpetual motion of a fluidic universe and for the flux of human life. The Sphinx is the equilibrium therein. The transliteration of Taro as Rota is inscribed on the wheel, counterchanged with the letters of the Divine Name--to shew that Providence is imphed through all. But this is the Divine intention within, and the similar intention without is exemplified by the four Living Creatures. Sometimes the sphinx is represented couchant on a pedestal above, which defrauds the symbolism by stultifying the essential idea of stability amidst movement.Behind the general notion expressed in the symbol there lies the denial of chance and the fatality which is implied therein. It may be added that, from the days of L\u00e9vi onward, the occult explanations of this card are--even for occultism itself--of a singularly fatuous kind. It has been said to mean principle, fecundity, virile honour, ruling authority, etc. The findings of common fortune-telling are better than this on their own plane.",
+    numeral: '10',
+  },
+  11: {
+    name: 'Justice',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Making an objective decision',
+      'Weighing an issue carefully before taking action',
+      'Appropriately scaling your reaction to a situation',
+      'Getting all the facts',
+      'Considering evidence',
+      'Deliberating',
+    ],
+    reversedMeanings: [
+      'Delivering harsh criticism',
+      'Obsessing on rules and regulations',
+      'Playing by the book even when it is destructive or counterproductive to do so',
+      'Confusing snap decisions with timely action',
+      'Playing favorites',
+    ],
+    suit: 'Trump',
+    keywords: ['balance', 'law', 'fairness', 'objectivity'],
+    img: 'm11.jpg',
+    symbolism: 'TODO',
+    description:
+      'As this card follows the traditional symbolism and carries above all its obvious meanings, there is little to say regarding it outside the few considerations collected in the first part, to which the reader is referred.It will be seen, however, that the figure is seated between pillars, like the High Priestess, and on this account it seems desirable to indicate that the moral principle which deals unto every man according to his works--while, of course, it is in strict analogy with higher things;--differs in its essence from the spiritual justice which is involved in the idea of election. The latter belongs to a mysterious order of Providence, in virtue of which it is possible for certain men to conceive the idea of dedication to the highest things. The operation of this is like the breathing of the Spirit where it wills, and we have no canon of criticism or ground of explanation concerning it. It is analogous to the possession of the fairy gifts and the high gifts and the gracious gifts of the poet: we have them or have not, and their presence is as much a mystery as their absence. The law of Justice is not however involved by either alternative. In conclusion, the pillars of Justice open into one world and the pillars of the High Priestess into another.',
+    numeral: '11',
+  },
+  12: {
+    name: 'The Hanged Man',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Seeing growth opportunities in unpleasant events',
+      'Experiencing a dramatic change in personal perspective',
+      'Making the best of an unforeseen change in your life or work',
+      'Suspending disbelief',
+      'Making sacrifices',
+    ],
+    reversedMeanings: [
+      'Being untrue to yourself and your values',
+      'Refusing to make sacrifices when appropriate',
+      'Refusing to adapt to new situations',
+      'Blaming others',
+      'Profiting at the expense of others',
+    ],
+    suit: 'Trump',
+    keywords: [
+      'enlightenment',
+      'sacrifice',
+      'perspective',
+      'suspension',
+      'reversals',
+    ],
+    img: 'm12.jpg',
+    symbolism: 'TODO',
+    description:
+      'The gallows from which he is suspended forms a Tau cross, while the figure--from the position of the legs--forms a fylfot cross. There is a nimbus about the head of the seeming martyr. It should be noted (1) that the tree of sacrifice is living wood, with leaves thereon; (2) that the face expresses deep entrancement, not suffering; (3) that the figure, as a whole, suggests life in suspension, but life and not death. It is a card of profound significance, but all the significance is veiled. One of his editors suggests that Éliphas Lévi did not know the meaning, which is unquestionable nor did the editor himself. It has been called falsely a card of martyrdom, a card a of prudence, a card of the Great Work, a card of duty; but we may exhaust all published interpretations and find only vanity. I will say very simply on my own part that it expresses the relation, in one of its aspects, between the Divine and the Universe.He who can understand that the story of his higher nature is imbedded in this symbolism will receive intimations concerning a great awakening that is possible, and will know that after the sacred Mystery of Death there is a glorious Mystery of Resurrection.',
+    numeral: '12',
+  },
+  13: {
+    name: 'Death',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Bringing an unpleasant phase of life to an end',
+      'Recognizing and celebrating the conclusion of something',
+      'Putting bad habits to rest',
+      'Becoming a new person',
+      'Leaving one person, place, or thing for another',
+      'Letting go',
+    ],
+    reversedMeanings: [
+      'Obsessing on death and dying',
+      'Refusing to give up old habits or unhealthy relationships',
+      'Insisting that everything and everyone should stay the same forever',
+      'Failing to take good care of yourself',
+    ],
+    suit: 'Trump',
+    keywords: ['ending', 'conclusion', 'transition', 'passage', 'departure'],
+    img: 'm13.jpg',
+    symbolism: 'TODO',
+    description:
+      'The veil or mask of life is perpetuated in change, transformation and passage from lower to higher, and this is more fitly represented in the rectified Tarot by one of the apocalyptic visions than by the crude notion of the reaping skeleton. Behind it lies the whole world of ascent in the spirit. The mysterious horseman moves slowly, bearing a black banner emblazoned with the Mystic Rose, which signifies life. Between two pillars on the verge of the horizon there shines the sun of immortality. The horseman carries no visible weapon, but king and child and maiden fall before him, while a prelate with clasped hands awaits his end.There should be no need to point out that the suggestion of death which I have made in connection with the previous card is, of course, to be understood mystically, but this is not the case in the present instance. The natural transit of man to the next stage of his being either is or may be one form of his progress, but the exotic and almost unknown entrance, while still in this life, into the state of mystical death is a change in the form of consciousness and the passage into a state to which ordinary death is neither the path nor gate. The existing occult explanations of the 13th card are, on the whole, better than usual, rebirth, creation, destination, renewal, and the rest.',
+    numeral: '13',
+  },
+  14: {
+    name: 'Temperance',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Bringing opposites together',
+      'Moderating your actions or emotions',
+      'Finding middle ground',
+      'Reaching compromises',
+      'Synthesizing solutions that please everyone involved',
+      'Using the old to make something new',
+    ],
+    reversedMeanings: [
+      'Going to extremes',
+      'Disrupting group efforts',
+      'Ignoring healthy approaches to life',
+      'Becoming an addict',
+      'Practicing gluttony',
+      'Tearing something or someone apart',
+      'Breaking alliances',
+    ],
+    suit: 'Trump',
+    keywords: ['blending', 'synthesis', 'mediation', 'combination', 'harmony'],
+    img: 'm14.jpg',
+    symbolism: 'TODO',
+    description:
+      'A winged angel, with the sign of the sun upon his forehead and on his breast the square and triangle of the septenary. I speak of him in the masculine sense, but the figure is neither male nor female. It is held to be pouring the essences of life from chalice to chalice. It has one foot upon the earth and one upon waters, thus illustrating the nature of the essences. A direct path goes up to certain heights on the verge of the horizon, and above there is a great light, through which a crown is seen vaguely. Hereof is some part of the Secret of Eternal Life, as it is possible to man in his incarnation. All the conventional emblems are renounced herein.So also are the conventional meanings, which refer to changes in the seasons, perpetual movement of life and even the combination of ideas. It is, moreover, untrue to say that the figure symbolizes the genius of the sun, though it is the analogy of solar light, realized in the third part of our human triplicity. It is called Temperance fantastically, because, when the rule of it obtains in our consciousness, it tempers, combines and harmonises the psychic and material natures. Under that rule we know in our rational part something of whence we came and whither we are going.',
+    numeral: '14',
+  },
+  15: {
+    name: 'The Devil',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Appreciating the luxuries that life has to offer',
+      'Being comfortable in your own skin',
+      'Enjoying your sexuality',
+      'Splurging on an expensive personal item',
+      'Embracing the fact that everyone has a darker side',
+      'Dealing with unhealthy impulses in healthy ways',
+    ],
+    reversedMeanings: [
+      'Putting excessive emphasis on appearances',
+      'Always wanting more',
+      'Valuing possessions more than people or relationships',
+      'Allowing base instincts to govern your life',
+      'Being selfish',
+      'Attributing your own dark impulses to outside forces or other people',
+    ],
+    suit: 'Trump',
+    keywords: ['shadow', 'materialism', 'bondage', 'delusion'],
+    img: 'm15.jpg',
+    symbolism: 'TODO',
+    description:
+      'The design is an accommodation, mean or harmony, between several motives mentioned in the first part. The Horned Goat of Mendes, with wings like those of a bat, is standing on an altar. At the pit of the stomach there is the sign of Mercury. The right hand is upraised and extended, being the reverse of that benediction which is given by the Hierophant in the fifth card. In the left hand there is a great flaming torch, inverted towards the earth. A reversed pentagram is on the forehead. There is a ring in front of the altar, from which two chains are carried to the necks of two figures, male and female. These are analogous with those of the fifth card, as if Adam and Eve after the Fall. Hereof is the chain and fatality of the material life.The figures are tailed, to signify the animal nature, but there is human intelligence in the faces, and he who is exalted above them is not to be their master for ever. Even now, he is also a bondsman, sustained by the evil that is in him and blind to the liberty of service. With more than his usual derision for the arts which he pretended to respect and interpret as a master therein, Éliphas Lévi affirms that the Baphometic figure is occult science and magic. Another commentator says that in the Divine world it signifies predestination, but there is no correspondence in that world with the things which below are of the brute. What it does signify is the Dweller on the Threshold without the Mystical Garden when those are driven forth therefrom who have eaten the forbidden fruit.',
+    numeral: '15',
+  },
+  16: {
+    name: 'The Tower',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Breaking out of old, confining habits and mindsets',
+      'Clearing the way for new growth',
+      'Dispelling the influence of an inflated ego',
+      'Getting back to basics',
+      'Stripping away harmful illusions',
+      'Receiving sudden insight',
+    ],
+    reversedMeanings: [
+      'Clinging to traditions that repress growth',
+      'Engaging in willful blindness',
+      'Rejecting evidence that change is needed',
+      'Ignoring guidance from a higher power',
+      'Maliciously engaging in destructive behavior',
+    ],
+    suit: 'Trump',
+    keywords: [
+      'demolition',
+      'upheaval',
+      'deconstruction',
+      'disaster',
+      'destruction',
+    ],
+    img: 'm16.jpg',
+    symbolism: 'TODO',
+    description:
+      'Occult explanations attached to this card are meagre and mostly disconcerting. It is idle to indicate that it depicts min in all its aspects, because it bears this evidence on the surface. It is said further that it contains the first allusion to a material building, but I do not conceive that the Tower is more or less material than the pillars which we have met with in three previous cases. I see nothing to warrant Papus in supposing that it is literally the fall of Adam, but there is more in favour of his alternative--that it signifies the materialization of the spiritual word. The bibliographer Christian imagines that it is the downfall of the mind, seeking to penetrate the mystery of God. I agree rather with Grand Orient that it is the ruin of the House of We, when evil has prevailed therein, and above all that it is the rending of a House of Doctrine. I understand that the reference is, however, to a House of Falsehood. It illustrates also in the most comprehensive way the old truth that "except the Lord build the house, they labour in vain that build it."There is a sense in which the catastrophe is a reflection from the previous card, but not on the side of the symbolism which I have tried to indicate therein. It is more correctly a question of analogy; one is concerned with the fall into the material and animal state, while the other signifies destruction on the intellectual side. The Tower has been spoken of as the chastisement of pride and the intellect overwhelmed in the attempt to penetrate the Mystery of God; but in neither case do these explanations account for the two persons who are the living sufferers. The one is the literal word made void and the other its false interpretation. In yet a deeper sense, it may signify also the end of a dispensation, but there is no possibility here for the consideration of this involved question.',
+    numeral: '16',
+  },
+  17: {
+    name: 'The Star',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Hoping for the best',
+      'Believing good things happen to good people',
+      'Seeing events in the best possible light',
+      'Adopting a generous spirit',
+      'Seeking guidance from above',
+      'Embracing possibility over probability',
+    ],
+    reversedMeanings: [
+      'Denying unpleasant truths',
+      'Denying personal accountability and saying, "Things just happen!"',
+      'Ignoring signs and omens',
+      'Preferring illusion to reality',
+      'Spreading pessimism and stinginess of spirit',
+    ],
+    suit: 'Trump',
+    keywords: [
+      'hope',
+      'optimism',
+      'openness',
+      'certainty',
+      'faith',
+      'longing',
+      'truth',
+    ],
+    img: 'm17.jpg',
+    symbolism: 'TODO',
+    description:
+      'A great, radiant star of eight rays, surrounded by seven lesser stars--also of eight rays. The female figure in the foreground is entirely naked. Her left knee is on the land and her right foot upon the water. She pours Water of Life from two great ewers, irrigating sea and land. Behind her is rising ground and on the right a shrub or tree, whereon a bird alights. The figure expresses eternal youth and beauty. The star is l\'étoile flamboyante, which appears in Masonic symbolism, but has been confused therein. That which the figure communicates to the living scene is the substance of the heavens and the elements. It has been said truly that the mottoes of this card are "Waters of Life freely" and "Gifts of the Spirit."The summary of several tawdry explanations says that it is a card of hope. On other planes it has been certified as immortality and interior light. For the majority of prepared minds, the figure will appear as the type of Truth unveiled, glorious in undying beauty, pouring on the waters of the soul some part and measure of her priceless possession. But she is in reality the Great Mother in the Kabalistic Sephira Binah, which is supernal Understanding, who communicates to the Sephiroth that are below in the measure that they can receive her influx.',
+    numeral: '17',
+  },
+  18: {
+    name: 'The Moon',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Enjoying healthy fantasies and daydreams',
+      'Using your imagination',
+      'Practicing magic or celebrating the magic of everyday life',
+      'Attuning yourself to the cycles of nature',
+      'Embracing the unknown',
+    ],
+    reversedMeanings: [
+      'Becoming unable to separate fantasy from reality',
+      'Suffering from delusions',
+      'Losing your appreciation for the fantastic or magical',
+      'Adopting a ruthlessly logical mindset',
+      "Failing to appreciate life's mysteries",
+    ],
+    suit: 'Trump',
+    keywords: ['mystery', 'fantasy', 'imagination', 'dreams', 'uncertainty'],
+    img: 'm18.jpg',
+    symbolism: 'TODO',
+    description:
+      'The distinction between this card and some of the conventional types is that the moon is increasing on what is called the side of mercy, to the right of the observer. It has sixteen chief and sixteen secondary rays. The card represents life of the imagination apart from life of the spirit. The path between the towers is the issue into the unknown. The dog and wolf are the fears of the natural mind in the presence of that place of exit, when there is only reflected light to guide it.The last reference is a key to another form of symbolism. The intellectual light is a reflection and beyond it is the unknown mystery which it cannot shew forth. It illuminates our animal nature, types of which are represented below--the dog, the wolf and that which comes up out of the deeps, the nameless and hideous tendency which is lower than the savage beast. It strives to attain manifestation, symbolized by crawling from the abyss of water to the land, but as a rule it sinks back whence it came. The face of the mind directs a calm gaze upon the unrest below; the dew of thought falls; the message is: Peace, be still; and it may be that there shall come a calm upon the animal nature, while the abyss beneath shall cease from giving up a form.',
+    numeral: '18',
+  },
+  19: {
+    name: 'The Sun',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Seeing things clearly',
+      'Experiencing intense joy',
+      'Celebrating your own successes',
+      "Knowing you're good at what you do",
+      'Gaining recognition for your personal genius',
+    ],
+    reversedMeanings: [
+      'Being dazzled by your own accomplishments',
+      'Becoming absorbed in your own self-image',
+      'Feeling rushed and distracted',
+      'Exerting yourself to the point of exhaustion',
+      'Overstating your abilities or misrepresenting your achievements',
+    ],
+    suit: 'Trump',
+    keywords: ['joy', 'brilliance', 'validation', 'attention', 'energy'],
+    img: 'm19.jpg',
+    symbolism: 'TODO',
+    description:
+      'The naked child mounted on a white horse and displaying a red standard has been mentioned already as the better symbolism connected with this card. It is the destiny of the Supernatural East and the great and holy light which goes before the endless procession of humanity, coming out from the walled garden of the sensitive life and passing on the journey home. The card signifies, therefore, the transit from the manifest light of this world, represented by the glorious sun of earth, to the light of the world to come, which goes before aspiration and is typified by the heart of a child.But the last allusion is again the key to a different form or aspect of the symbolism. The sun is that of consciousness in the spirit - the direct as the antithesis of the reflected light. The characteristic type of humanity has become a little child therein--a child in the sense of simplicity and innocence in the sense of wisdom. In that simplicity, he bears the seal of Nature and of Art; in that innocence, he signifies the restored world. When the self-knowing spirit has dawned in the consciousness above the natural mind, that mind in its renewal leads forth the animal nature in a state of perfect conformity.',
+    numeral: '19',
+  },
+  20: {
+    name: 'Judgment',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Receiving a wake-up call',
+      'Discovering a new purpose in life',
+      'Becoming totally and completely yourself',
+      'Receiving a well-deserved reward',
+      'Passing an evaluation or examination',
+      'Welcoming the start of a new phase of life',
+    ],
+    reversedMeanings: [
+      'Being weighed in the balances and found wanting',
+      'Failing to measure up to a well-defined standard',
+      'Being caught goofing off or misbehaving',
+      'Failing to prepare for an examination you know is coming',
+      'Rejecting an opportunity to reinvent yourself',
+    ],
+    suit: 'Trump',
+    keywords: [
+      'revival',
+      'renewal',
+      'resurrection',
+      'evaluation',
+      'invitation',
+    ],
+    img: 'm20.jpg',
+    symbolism: 'TODO',
+    description:
+      'I have said that this symbol is essentially invariable in all Tarot sets, or at least the variations do not alter its character. The great angel is here encompassed by clouds, but he blows his bannered trumpet, and the cross as usual is displayed on the banner. The dead are rising from their tombs--a woman on the right, a man on the left hand, and between them their child, whose back is turned. But in this card there are more than three who are restored, and it has been thought worth while to make this variation as illustrating the insufficiency of current explanations. It should be noted that all the figures are as one in the wonder, adoration and ecstacy expressed by their attitudes. It is the card which registers the accomplishment of the great work of transformation in answer to the summons of the Supernal--which summons is heard and answered from within.Herein is the intimation of a significance which cannot well be carried further in the present place. What is that within us which does sound a trumpet and all that is lower in our nature rises in response--almost in a moment, almost in the twinkling of an eye? Let the card continue to depict, for those who can see no further, the Last judgment and the resurrection in the natural body; but let those who have inward eyes look and discover therewith. They will understand that it has been called truly in the past a card of eternal life, and for this reason it may be compared with that which passes under the name of Temperance.',
+    numeral: '20',
+  },
+  21: {
+    name: 'The World',
+    arcana: 'Major Arcana',
+    uprightMeanings: [
+      'Having it all',
+      'Knowing and loving yourself as completely as possible',
+      'Seeing the interconnection of all things and people',
+      'Enhancing your perspective',
+      'Living life to its fullest',
+      'Understanding the meaning of life',
+    ],
+    reversedMeanings: [
+      'Allowing greed and envy to prevent you from enjoying what you do possess',
+      'Failing to see the larger design in ordinary events',
+      'Believing that everything that exists can be touched, counted, or measured',
+      'Failing to see the divine reflected in those around you',
+    ],
+    suit: 'Trump',
+    keywords: [
+      'wholeness',
+      'integration',
+      'totality',
+      'completeness',
+      'fullness',
+    ],
+    img: 'm21.jpg',
+    symbolism: 'TODO',
+    description:
+      'As this final message of the Major Trumps is unchanged--and indeed unchangeable--in respect of its design, it has been partly described already regarding its deeper sense. It represents also the perfection and end of the Cosmos, the secret which is within it, the rapture of the universe when it understands itself in God. It is further the state of the soul in the consciousness of Divine Vision, reflected from the self-knowing spirit. But these meanings are without prejudice to that which I have said concerning it on the material side.It has more than one message on the macrocosmic side and is, for example, the state of the restored world when the law of manifestation shall have been carried to the highest degree of natural perfection. But it is perhaps more especially a story of the past, referring to that day when all was declared to be good, when the morning stars sang together and all the Sons of God shouted for joy. One of the worst explanations concerning it is that the figure symbolizes the Magus when he has reached the highest degree of initiation; another account says that it represents the absolute, which is ridiculous. The figure has been said to stand for Truth, which is, however, more properly allocated to the seventeenth card. Lastly, it has been called the Crown of the Magi.',
+    numeral: '21',
+  },
+  22: {
+    name: 'Ace of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Trusting your feelings',
+      'Opening yourself to spirit',
+      'Accepting and returning affection',
+      'Getting in touch with what motivates you',
+      'Taking advantage of an opportunity to express love to others',
+      'Listening to the still, small voice',
+    ],
+    reversedMeanings: [
+      'Hiding your feelings',
+      'Spurning an opportunity to love or be loved',
+      'Numbing yourself to spiritual yearnings',
+      'Rejecting the counsel of your heart',
+      'Becoming a puppet of your own emotions',
+      'Indulging in hysteria or obsession',
+    ],
+    suit: 'Cups',
+    keywords: ['intuition', 'spirituality', 'affection', 'motivation'],
+    img: 'c01.jpg',
+    symbolism: 'TODO',
+    description:
+      'The waters are beneath, and thereon are water-lilies; the hand issues from the cloud, holding in its palm the cup, from which four streams are pouring; a dove, bearing in its bill a cross-marked Host, descends to place the Wafer in the Cup; the dew of water is falling on all sides. It is an intimation of that which may lie behind the Lesser Arcana.',
+    numeral: '1',
+  },
+  23: {
+    name: 'Two of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Being drawn to someone',
+      'Longing for someone or something',
+      'Acting on your desires',
+      'Discovering a feeling is mutual',
+      'Doing what makes you feel good',
+      'Merging',
+      'Healing broken ties',
+      'Admitting two people feel differently about each other and moving on',
+    ],
+    reversedMeanings: [
+      'Burning bridges',
+      'Becoming caught up in unhealthy codependency',
+      'Shutting out anyone but your chosen few',
+      'Obsessing on someone who does not return your affections',
+      'Despairing over finding "The One"',
+      'Deceiving yourself about your true orientation',
+    ],
+    suit: 'Cups',
+    keywords: ['union', 'attraction', 'combination', 'affection'],
+    img: 'c02.jpg',
+    symbolism: 'TODO',
+    description:
+      "A youth and maiden are pledging one another, and above their cups rises the Caduceus of Hermes, between the great wings of which there appears a lion's head. It is a variant of a sign which is found in a few old examples of this card. Some curious emblematical meanings are attached to it, but they do not concern us in this place.",
+    numeral: '2',
+  },
+  24: {
+    name: 'Three of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Celebrating your feelings or connections with others',
+      'Expressing joy through song, dance, or physical affection',
+      'Working together with others who share your feelings',
+      'Performing acts of service as a way of saying, "I love you"',
+      'Embracing unconventional romantic arrangements',
+    ],
+    reversedMeanings: [
+      'Mistaking giddiness for true affection',
+      'Being dominated by manic emotions',
+      'Expecting everyone to always feel the same way you do',
+      'Demanding unreasonable support from friends or family',
+      'Partying to a dangerous or unhealthy extent',
+    ],
+    suit: 'Cups',
+    keywords: ['celebration', 'expression', 'community', 'friendliness'],
+    img: 'c03.jpg',
+    symbolism: 'TODO',
+    description:
+      'Maidens in a garden-ground with cups uplifted, as if pledging one another.',
+    numeral: '3',
+  },
+  25: {
+    name: 'Four of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Maintaining your emotional stability',
+      'Refusing to give in to overwhelming emotions',
+      'Appreciating what you have and refusing to take it for granted',
+      'Seeing the value of long-term commitments',
+    ],
+    reversedMeanings: [
+      'Being bored',
+      'Daydreaming at the expense of your work',
+      'Refusing to be engaged by opportunity',
+      'Taking people and relationships for granted',
+      'Ignoring romantic or spiritual opportunities',
+      'Spurning inspiration',
+      'Feeling everything should stay "just like it is"',
+    ],
+    suit: 'Cups',
+    keywords: [
+      'boredom',
+      'listlessness',
+      'lethargy',
+      'stability',
+      'ingratitude',
+    ],
+    img: 'c04.jpg',
+    symbolism: 'TODO',
+    description:
+      'A young man is seated under a tree and contemplates three cups set on the grass before him; an arm issuing from a cloud offers him another cup. His expression notwithstanding is one of discontent with his environment.',
+    numeral: '4',
+  },
+  26: {
+    name: 'Five of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Acknowledging loss and moving on',
+      'Focusing on how the glass remains "half-full"',
+      'Finding the silver lining in a dark cloud',
+      'Recognizing that loss is a natural part of life',
+      'Embracing healthy grief',
+      'Learning lessons from harsh consequences',
+    ],
+    reversedMeanings: [
+      'Wallowing in unhealthy grief or self-pity',
+      'Refusing to move on and let go',
+      'Clinging to the past',
+      'Obsessing on past lives and past loves',
+      'Failing to live in the present',
+      'Beating yourself up over past mistakes',
+      'Allowing fear of failure to limit your efforts',
+    ],
+    suit: 'Cups',
+    keywords: [
+      'loss',
+      'despair',
+      're-evaluation',
+      'regret',
+      'uncertainty',
+      'repentance',
+    ],
+    img: 'c05.jpg',
+    symbolism: 'TODO',
+    description:
+      'A dark, cloaked figure, looking sideways at three prone cups two others stand upright behind him; a bridge is in the background, leading to a small keep or holding. Divanatory Meanings: It is a card of loss, but something remains over; three have been taken, but two are left; it is a card of inheritance, patrimony, transmission, but not corresponding to expectations; with some interpreters it is a card of marriage, but not without bitterness or frustration. Reversed: News, alliances, affinity, consanguinity, ancestry, return, false projects',
+    numeral: '5',
+  },
+  27: {
+    name: 'Six of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Donating your time and talents to others',
+      'Taking satisfaction in knowing how your efforts will aid others',
+      'Creating a "win-win" scenario',
+      'Giving even when you know repayment is not possible',
+      'Being motivated to do a good deed',
+    ],
+    reversedMeanings: [
+      'Linking your sense of self-worth to the appraisals of others',
+      'Striving to appear more needy than you really are',
+      'Taking undeserved or unmerited charity',
+      'Bragging about your charitable efforts',
+      'Profiteering in times of distress',
+      'Refusing to share a burden',
+    ],
+    suit: 'Cups',
+    keywords: ['charity', 'sharing', 'sacrifice', 'cooperation', 'fairness'],
+    img: 'c06.jpg',
+    symbolism: 'TODO',
+    description: 'Children in an old garden, their cups filled with flowers.',
+    numeral: '6',
+  },
+  28: {
+    name: 'Seven of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Motivating yourself with images of future success',
+      'Using visualization to encourage progress',
+      'Taking an imaginative or creative approach to problem solving',
+      'Making dreams come true',
+      'Gleaning insight from personal visions',
+    ],
+    reversedMeanings: [
+      'Obsessing on imaginary fears or uncertain consequences',
+      'Giving in to emotional or political terrorism',
+      'Spending more time dreaming than working',
+      'Failing to envision the possible repercussions of your choices',
+      'Being controlled by fear',
+    ],
+    suit: 'Cups',
+    keywords: ['imagination', 'dreams', 'illusions', 'goals'],
+    img: 'c07.jpg',
+    symbolism: 'TODO',
+    description:
+      'Strange chalices of vision, but the images are more especially those of the fantastic spirit.',
+    numeral: '7',
+  },
+  29: {
+    name: 'Eight of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Wanting something better',
+      'Blazing your own trail',
+      'Realizing there must be more to life',
+      'Leaving an unhealthy situation behind',
+      'Starting your own business',
+      'Going on a retreat',
+      'Seeking the "still, small voice"',
+    ],
+    reversedMeanings: [
+      'Being implacable',
+      'Finding fault',
+      'Nitpicking',
+      'Refusing to settle down',
+      'Running away from problems or confrontations',
+      'Saying, "It\'s my way or the highway!"',
+      'Harping on past mistakes and disappointments',
+      'Threatening to quit as a strategy to get your way',
+    ],
+    suit: 'Cups',
+    keywords: [
+      'longing',
+      'dissatisfaction',
+      'quest',
+      'departure',
+      'withdrawal',
+    ],
+    img: 'c08.jpg',
+    symbolism: 'TODO',
+    description:
+      'A man of dejected aspect is deserting the cups of his felicity, enterprise, undertaking or previous concern.',
+    numeral: '8',
+  },
+  30: {
+    name: 'Nine of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Being delighted with your own achievements',
+      'Recognizing your own talents and abilities',
+      'Reveling in the good things life has to offer',
+      'Indulging yourself',
+      'Relaxing and unwinding',
+      'Having everything you need in order to feel complete',
+    ],
+    reversedMeanings: [
+      'Being smug',
+      'Satisfying yourself at the expense of others',
+      'Being selfish',
+      'Over-indulging',
+      'Avoiding work that needs to be done',
+      'Claiming achievements or skills you do not possess',
+      'Never being satisfied, no matter how much you have',
+    ],
+    suit: 'Cups',
+    keywords: ['satisfaction', 'sensuality', 'luxury', 'pleasure'],
+    img: 'c09.jpg',
+    symbolism: 'TODO',
+    description:
+      "A goodly personage has feasted to his heart's content, and abundant refreshment of wine is on the arched counter behind him, seeming to indicate that the future is also assured. The picture offers the material side only, but there are other aspects.",
+    numeral: '9',
+  },
+  31: {
+    name: 'Ten of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Having more than you ever dreamed',
+      "Being deeply thankful for all you've been given",
+      'Recognizing the Hand of God in the gifts the Universe brings your way',
+      'Experiencing transcendent joy',
+      'Achieving domestic bliss',
+    ],
+    reversedMeanings: [
+      'Comparing your achievements or relationships to unrealistic fantasy standards',
+      'Experiencing emotions so intense they blunt your ability to cope with reality',
+      'Feeling overwhelmed',
+      'Envying the achievements and happiness of others',
+    ],
+    suit: 'Cups',
+    keywords: ['joy', 'fulfillment', 'overwhelming emotion', 'giddiness'],
+    img: 'c10.jpg',
+    symbolism: 'TODO',
+    description:
+      'Appearance of Cups in a rainbow; it is contemplated in wonder and ecstacy by a man and woman below, evidently husband and wife. His right arm is about her; his left is raised upward; she raises her right arm. The two children dancing near them have not observed the prodigy but are happy after their own manner. There is a home-scene beyond.',
+    numeral: '10',
+  },
+  32: {
+    name: 'Page of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Showing your emotions freely',
+      'Throwing yourself into romance',
+      'Nursing a secret crush',
+      'Indulging in romantic fantasy',
+      'Starting a new relationship',
+      'Recalling your first love',
+      'Experiencing love for the first time',
+      'Converting to a new religion',
+    ],
+    reversedMeanings: [
+      'Mistaking a crush for true love',
+      'Reading romantic intention into innocent action',
+      'Frantically trying to impress others',
+      'Indulging in overly-sweet sentimentality',
+      'Pretending to more romantic or spiritual experience than you possess',
+    ],
+    suit: 'Cups',
+    keywords: [
+      'enthusiasm',
+      'first impressions',
+      'romanticism',
+      'superficiality',
+    ],
+    img: 'c11.jpg',
+    symbolism: 'TODO',
+    description:
+      'A fair, pleasing, somewhat effeminate page, of studious and intent aspect, contemplates a fish rising from a cup to look at him. It is the pictures of the mind taking form.',
+    numeral: '11',
+  },
+  33: {
+    name: 'Knight of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Being deeply committed to a cause',
+      'Giving in to strong emotions, from excitement to depression',
+      'Acting on intuition alone',
+      'Solving problems intuitively',
+      'Believing in and basing decisions on ideals instead of realities',
+      'Bringing intuition or passion to the table',
+    ],
+    reversedMeanings: [
+      'Becoming a fanatic',
+      'Rejecting information that suggests your intuitions are misguided',
+      'Allowing your emotions to control you',
+      'Giving in to jealousy, confrontation, and peer pressure',
+      'Hiding or ignoring intuitive insights',
+    ],
+    suit: 'Cups',
+    keywords: ['fervor', 'zeal', 'moodiness', 'illumination'],
+    img: 'c12.jpg',
+    symbolism: 'TODO',
+    description:
+      'Graceful, but not warlike; riding quietly, wearing a winged helmet, referring to those higher graces of the imagination which sometimes characterize this card. He too is a dreamer, but the images of the side of sense haunt him in his vision.',
+    numeral: '12',
+  },
+  34: {
+    name: 'Queen of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Allowing yourself to be moved by the plight of others',
+      'Feeling strong emotions',
+      'Possessing unusual sympathy or empathy',
+      'Trusting your feelings to guide you',
+      'Calling on psychic abilities',
+      'Achieving unity with Spirit',
+    ],
+    reversedMeanings: [
+      'Becoming so caught up in matters of Spirit, you become detached from the world',
+      'Allowing empathy to disable you (instead of inspire action)',
+      'Using psychic abilities to wield covert influence',
+      'Wallowing in emotionalism, sentiment, or self-pity',
+    ],
+    suit: 'Cups',
+    keywords: [
+      'insightfulness',
+      'spirituality',
+      'compassion',
+      'empathy',
+      'instinct',
+    ],
+    img: 'c13.jpg',
+    symbolism: 'TODO',
+    description:
+      'Beautiful, fair, dreamy--as one who sees visions in a cup. This is, however, only one of her aspects; she sees, but she also acts, and her activity feeds her dream.',
+    numeral: '13',
+  },
+  35: {
+    name: 'King of Cups',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Keeping a stiff upper lip',
+      'Being brave and clear in the face of adverse circumstances',
+      'Sharing experience as a way of comforting others',
+      'Making fair and empathetic decisions',
+      'Honoring the spirit, not just the letter, of the law',
+    ],
+    reversedMeanings: [
+      'Allowing yourself to become rigid and unemotional',
+      'Making unfair decisions based on a hidden agenda',
+      'Making decisions without regard for their emotional impact on others',
+      'Abusing spiritual authority',
+      'Using emotional or spiritual leverage to exercise unhealthy control over others',
+    ],
+    suit: 'Cups',
+    keywords: ['wisdom', 'diplomacy', 'restraint', 'composure'],
+    img: 'c14.jpg',
+    symbolism: 'TODO',
+    description:
+      'He holds a short sceptre in his left hand and a great cup in his right; his throne is set upon the sea; on one side a ship is riding and on the other a dolphin is leaping. The implicit is that the Sign of the Cup naturally refers to water, which appears in all the court cards.',
+    numeral: '14',
+  },
+  36: {
+    name: 'Ace of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Making objective decisions',
+      'Applying logic',
+      'Reasoning your way out of a difficult situation',
+      'Solving puzzles',
+      'Thinking things through',
+      'Emphasizing the facts',
+      'Clearing your mind',
+      'Seeking clarity',
+    ],
+    reversedMeanings: [
+      'Applying ruthless or twisted logic',
+      'Gloating over your own superior intellect',
+      'Using quick thinking to deceive or confuse others',
+      'Confusing snap judgments with quick thinking',
+      'Making decisions without thinking through consequences',
+    ],
+    suit: 'Swords',
+    keywords: ['logic', 'objectivity', 'intellect', 'choice'],
+    img: 's01.jpg',
+    symbolism: 'TODO',
+    description:
+      'A hand issues from a cloud, grasping as word, the point of which is encircled by a crown.',
+    numeral: '1',
+  },
+  37: {
+    name: 'Two of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Refusing to make a decision without getting the facts',
+      'Exploring both sides of an argument',
+      'Arguing passionately for what you believe in',
+      'Weighing the issues',
+      'Encouraging the open exchange of ideas',
+      'Discussing political or religious issues without getting "hot under the collar"',
+    ],
+    reversedMeanings: [
+      'Rejecting evidence that conflicts with dearly-held beliefs',
+      'Arguing with others just for the sake of doing so',
+      'Nit-picking',
+      "Putting off a decision because you're afraid to face the consequences",
+      'Preventing others from getting the information they need to make good decisions',
+    ],
+    suit: 'Swords',
+    keywords: ['denial', 'debate', 'impasse', 'truce'],
+    img: 's02.jpg',
+    symbolism: 'TODO',
+    description:
+      'A hoodwinked female figure balances two swords upon her shoulders.',
+    numeral: '2',
+  },
+  38: {
+    name: 'Three of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Being brave enough to see things as they really are',
+      'Exercising your critical eye',
+      'Being your own best critic',
+      "Acknowledging that things don't always turn out as planned",
+      'Moving past heartbreak to embrace a painful truth',
+    ],
+    reversedMeanings: [
+      'Wallowing in despair',
+      'Allowing yourself to be completely crushed by the thoughts, words, or deeds of another',
+      'Judging yourself too harshly',
+      'Holding yourself to an unrealistic standard of excellence',
+      'Wearing your heart on your sleeve while carrying a chip on your shoulder',
+    ],
+    suit: 'Swords',
+    keywords: [
+      'variance',
+      'difference',
+      'dissatisfaction',
+      'heartache',
+      'rejection',
+    ],
+    img: 's03.jpg',
+    symbolism: 'TODO',
+    description: 'Three swords piercing a heart; cloud and rain behind.',
+    numeral: '3',
+  },
+  39: {
+    name: 'Four of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Thinking over your plans before putting them into action',
+      'Pausing to meditate or clear your mind',
+      'Taking time to understand someone or something before criticizing it',
+      'Resting',
+      'Occupying your thoughts with a healthy distraction',
+    ],
+    reversedMeanings: [
+      'Failing to think things through',
+      'Mistaking procrastination for thoughtfulness',
+      'Adopting a point of view and refusing to reconsider your conclusions, even when presented with refuting evidence',
+      'Allowing chaos and whimsy to dominate your thoughts',
+    ],
+    suit: 'Swords',
+    keywords: ['meditation', 'contemplation', 'perspective', 'mindset'],
+    img: 's04.jpg',
+    symbolism: 'TODO',
+    description:
+      'The effigy of a knight in the attitude of prayer, at full length upon his tomb.',
+    numeral: '4',
+  },
+  40: {
+    name: 'Five of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Acting in your own best interest',
+      'Choosing to stand up for yourself',
+      'Not backing down from disagreement and discord',
+      'Taking a stand',
+      'Refusing to go along with an unethical plan',
+      'Knowing when to bend the rules',
+    ],
+    reversedMeanings: [
+      'Taking advantage of others',
+      'Intimidating others',
+      'Acting in an unethical manner',
+      'Picking fights',
+      'Using words to goad others into violence and irrationality',
+      "Ignoring rules you've agreed to abide by",
+      'Looking out for yourself while allowing harm to come to others',
+      'Gloating over victory',
+    ],
+    suit: 'Swords',
+    keywords: [
+      'selfishness',
+      'hostility',
+      'irrationality',
+      'self-preservation',
+    ],
+    img: 's05.jpg',
+    symbolism: 'TODO',
+    description:
+      'A disdainful man looks after two retreating and dejected figures. Their swords lie upon the ground. He carries two others on his left shoulder, and a third sword is in his right hand, point to earth. He is the master in possession of the field.',
+    numeral: '5',
+  },
+  41: {
+    name: 'Six of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Making the best of a bad situation',
+      'Recovering from defeat',
+      'Resetting expectations',
+      'Making allowances for unexpected circumstances',
+      'Helping others who find themselves in dire circumstances',
+      'Changing the way you see the world',
+      'Broadening your perspective through study or travel',
+    ],
+    reversedMeanings: [
+      'Refusing to accept that things have changed',
+      'Playing the victim',
+      'Rejecting the idea that your actions have consequences',
+      'Applying scientific criteria to matters of faith, or confusing faith with science',
+      'Believing the whole world should be like your small corner of it',
+    ],
+    suit: 'Swords',
+    keywords: ['adaptation', 'adjustments', 'science', 'travel'],
+    img: 's06.jpg',
+    symbolism: 'TODO',
+    description:
+      'A ferryman carrying passengers in his punt to the further shore. The course is smooth, and seeing that the freight is light, it may be noted that the work is not beyond his strength.',
+    numeral: '6',
+  },
+  42: {
+    name: 'Seven of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      "Refusing to do something dishonest, even when there's no chance of ever being caught",
+      'Handling a difficult situation with finesse',
+      'Pointing out assumptions',
+      'Acting ethically in public and in private',
+      'Living a life that is beyond reproach',
+    ],
+    reversedMeanings: [
+      'Stealing or lying',
+      'Doing whatever you can get away with, simply because you can',
+      'Looking for a way around consequences',
+      'Justifying wicked behavior by focusing on the wickedness of others',
+      'Failing to examine your own motives and prejudices',
+    ],
+    suit: 'Swords',
+    keywords: ['dishonesty', 'presumption', 'sneakiness', 'assumptions'],
+    img: 's07.jpg',
+    symbolism: 'TODO',
+    description:
+      'A man in the act of carrying away five swords rapidly; the two others of the card remain stuck in the ground. A camp is close at hand.',
+    numeral: '7',
+  },
+  43: {
+    name: 'Eight of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Honoring limits',
+      'Respecting the rules',
+      "Deciding to go on a diet for your health's sake",
+      'Recognizing you cannot always be in control',
+      'Identifying obstacles to further progress',
+      'Refusing to think about unhealthy or unethical options',
+      'Asking for assistance',
+    ],
+    reversedMeanings: [
+      'Feeling trapped',
+      'Being lost in a maze of rules and regulations',
+      'Giving in to despair',
+      'Playing the victim',
+      'Allowing others to dictate what you can and cannot do',
+      'Being rendered helpless',
+      'Having very few options',
+      'Failing to look for a way out',
+    ],
+    suit: 'Swords',
+    keywords: ['restriction', 'limitation', 'confinement', 'helplessness'],
+    img: 's08.jpg',
+    symbolism: 'TODO',
+    description:
+      'A woman, bound and hoodwinked, with the swords of the card about her. Yet it is rather a card of temporary durance than of irretrievable bondage.',
+    numeral: '8',
+  },
+  44: {
+    name: 'Nine of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Refusing to worry about what you cannot control',
+      'Rejecting anxiety',
+      'Judging your own performance with kindness and gentleness',
+      'Using meditation to quiet a troubled mind',
+      'Confronting nightmares and fears',
+      'Drawing a conclusion and putting an issue out of your mind',
+    ],
+    reversedMeanings: [
+      'Torturing yourself with regrets',
+      'Second-guessing your every move',
+      'Beating yourself up for your mistakes',
+      'Depression',
+      'Obsessing on errors and overlooked details',
+      'Refusing to handle stress in healthy ways',
+      'Ruining your ability to appreciate the present by dwelling on the past',
+      'Debating irreversible decisions',
+    ],
+    suit: 'Swords',
+    keywords: ['remorse', 'worry', 'distraught', 'conclusion'],
+    img: 's09.jpg',
+    symbolism: 'TODO',
+    description:
+      'One seated on her couch in lamentation, with the swords over her. She is as one who knows no sorrow which is like unto hers. It is a card of utter desolation.',
+    numeral: '9',
+  },
+  45: {
+    name: 'Ten of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      "Seeing the signs that you've reached your limits",
+      'Paying attention to what your body is trying to tell you',
+      'Giving in to the need for rest and renewal',
+      "Acknowledging that you've hit bottom",
+      'Committing to a turnaround',
+      'Knowing the worst is over',
+    ],
+    reversedMeanings: [
+      'Accepting defeat prematurely',
+      'Driving yourself to total exhaustion, especially mentally',
+      'Experiencing a mental breakdown',
+      'Obsessing on a problem to the breaking point',
+      'Giving up',
+      'Refusing to move from thought to action',
+      'Deeply unhealthy thoughts',
+    ],
+    suit: 'Swords',
+    keywords: ['exhaustion', 'ruin', 'disaster', 'stamina', 'obsession'],
+    img: 's10.jpg',
+    symbolism: 'TODO',
+    description:
+      'A prostrate figure, pierced by all the swords belonging to the card.',
+    numeral: '10',
+  },
+  46: {
+    name: 'Page of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Pursuing a course of study',
+      'Asking good questions',
+      'Investing time in study and practice',
+      'Doing research',
+      'Making a habit of learning new things',
+      'Starting an investigation',
+      'Outlining what you need to know',
+      'Finding a mentor or teacher',
+    ],
+    reversedMeanings: [
+      'Pretending to knowledge or sophistication you do not possess',
+      'Cheating on an exam',
+      'Feigning interest as a way of gaining favor',
+      "Considering only the evidence that supports conclusions you've already drawn",
+      'Rejecting the wise counsel of experienced teachers',
+    ],
+    suit: 'Swords',
+    keywords: ['student', 'apprentice', 'scholarship', 'information'],
+    img: 's11.jpg',
+    symbolism: 'TODO',
+    description:
+      'A lithe, active figure holds a sword upright in both hands, while in the act of swift walking. He is passing over rugged land, and about his way the clouds are collocated wildly. He is alert and lithe, looking this way and that, as if an expected enemy might appear at any moment.',
+    numeral: '11',
+  },
+  47: {
+    name: 'Knight of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Speaking your mind',
+      'Making your opinions known',
+      'Offering constructive criticism',
+      'Sharing your knowledge',
+      'Making insightful observations',
+      'Pinpointing the problem',
+      'Clarifying what others have said',
+      'Giving clear direction to others',
+      'Uncovering the truth',
+    ],
+    reversedMeanings: [
+      'Stating your opinions as fact',
+      'Picking fights',
+      'Starting arguments',
+      'Using clever insults to undermine the confidence of others',
+      'Tossing reason out the window',
+      'Speaking without taking the feelings of others into account',
+      'Going on a witch hunt',
+      'Distorting evidence',
+    ],
+    suit: 'Swords',
+    keywords: ['bluntness', 'intelligence', 'incisiveness', 'investigation'],
+    img: 's12.jpg',
+    symbolism: 'TODO',
+    description:
+      'He is riding in full course, as if scattering his enemies. In the design he is really a prototypical hero of romantic chivalry. He might almost be Galahad, whose sword is swift and sure because he is clean of heart.',
+    numeral: '12',
+  },
+  48: {
+    name: 'Queen of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Exercising tact or using diplomacy',
+      'Defusing a tense situation',
+      'Knowing what to say and how to say it',
+      'Making others feel comfortable and confident',
+      'Bringing out the best in everyone',
+      'Having a way with words',
+      'Telling jokes',
+      'Possessing a knack for music, math, art, or science',
+    ],
+    reversedMeanings: [
+      'Knowing exactly what to say to destroy another person',
+      'Withholding critical information',
+      'Using a barbed tongue to upset others',
+      'Employing sarcasm',
+      'Mimicking others unkindly',
+      'Making light of the less fortunate',
+      'Being disrespectful',
+      "Failing to use the talent you've been given",
+    ],
+    suit: 'Swords',
+    keywords: ['grace', 'skill', 'wit', 'charm', 'aptitude'],
+    img: 's13.jpg',
+    symbolism: 'TODO',
+    description:
+      'Her right hand raises the weapon vertically and the hilt rests on an arm of her royal chair the left hand is extended, the arm raised her countenance is severe but chastened; it suggests familiarity with sorrow. It does not represent mercy, and, her sword notwithstanding, she is scarcely a symbol of power.',
+    numeral: '13',
+  },
+  49: {
+    name: 'King of Swords',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Expressing yourself with firmness and authority',
+      'Rendering a final decision',
+      'Consulting an expert',
+      'Calling in advisors and consultants',
+      'Coming to a final conclusion',
+      'Reaching a beneficial agreement based on sound information',
+    ],
+    reversedMeanings: [
+      'Insisting on having the last word',
+      'Flaunting your intellectual capability',
+      'Talking "over the heads" of others',
+      'Waffling on an important decision',
+      'Constantly changing your mind',
+      'Refusing to make choices that are in your own best interest',
+      "Wishing in vain you could take back what's been said",
+    ],
+    suit: 'Swords',
+    keywords: ['genius', 'expertise', 'decision', 'verdict'],
+    img: 's14.jpg',
+    symbolism: 'TODO',
+    description:
+      'He sits in judgment, holding the unsheathed sign of his suit. He recalls, of course, the conventional Symbol of justice in the Trumps Major, and he may represent this virtue, but he is rather the power of life and death, in virtue of his office.',
+    numeral: '14',
+  },
+  50: {
+    name: 'Ace of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Being inspired',
+      'Identifying an important goal',
+      'Being given the opportunity to do whatever you want to do',
+      'Giving or receiving direction',
+      'Seeing a solution',
+      'Creating something new',
+      'Being aroused, sexually or creatively',
+    ],
+    reversedMeanings: [
+      'Failing to take advantage of a great opportunity',
+      'Being ineffectual or lazy',
+      'Making an inadequate effort',
+      'Working toward a goal, but lacking the resources or initiative to achieve success',
+      'Setting inappropriate goals',
+      'Failing to take a stand',
+    ],
+    suit: 'Wands',
+    keywords: ['desire', 'inspiration', 'vision', 'creation', 'invention'],
+    img: 'w01.jpg',
+    symbolism: 'TODO',
+    description: 'A hand issuing from a cloud grasps a stout wand or club.',
+    numeral: '1',
+  },
+  51: {
+    name: 'Two of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Having a choice',
+      'Offering or being offered an option',
+      "Seeing the value of another person's approach",
+      'Understanding there\'s more than one way to "skin a cat"',
+      'Successfully doing more than one thing at a time',
+      'Being empowered to make a choice',
+    ],
+    reversedMeanings: [
+      'Misrepresenting your intentions',
+      'Doing one thing while desiring another',
+      'Changing course mid-stream for no good reason',
+      'Refusing to change your goal even when pursuing it no longer makes sense',
+      'Disregarding the input of others',
+    ],
+    suit: 'Wands',
+    keywords: ['conflict', 'decision', 'option', 'individuality'],
+    img: 'w02.jpg',
+    symbolism: 'TODO',
+    description:
+      'A tall man looks from a battlemented roof over sea and shore; he holds a globe in his right hand, while a staff in his left rests on the battlement; another is fixed in a ring. The Rose and Cross and Lily should be noticed on the left side.',
+    numeral: '2',
+  },
+  52: {
+    name: 'Three of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Putting a plan into motion',
+      'Taking that critical first step',
+      'Making good things happen',
+      'Going beyond your limits',
+      'Blazing new trails',
+      'Hitting the ground running',
+      'Seeing your plans come to fruition',
+    ],
+    reversedMeanings: [
+      'Procrastinating',
+      'Knowing what to do, but refusing to do it',
+      'Launching a project without a clear definition of who should do what',
+      'Rejecting an opportunity to try something new',
+      'Failing to finish what you start',
+    ],
+    suit: 'Wands',
+    keywords: ['implementation', 'action', 'exploration'],
+    img: 'w03.jpg',
+    symbolism: 'TODO',
+    description:
+      "A calm, stately personage, with his back turned, looking from a cliff's edge at ships passing over the sea. Three staves are planted in the ground, and he leans slightly on one of them.",
+    numeral: '3',
+  },
+  53: {
+    name: 'Four of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Sharing in a great celebration',
+      'Sharing in a communal sense of achievement and success',
+      'Preparing for a party',
+      'Working together toward a common goal',
+      'Giving or winning awards',
+    ],
+    reversedMeanings: [
+      'Keeping your nose to the grindstone',
+      'Recognizing good work by demanding more work',
+      'Failing to share in a group celebration',
+      'Allowing sour grapes to poison your moment in the sun',
+      'Refusing to do your part',
+    ],
+    suit: 'Wands',
+    keywords: [
+      'celebration',
+      'jubilation',
+      'community',
+      'teamwork',
+      'completion',
+    ],
+    img: 'w04.jpg',
+    symbolism: 'TODO',
+    description:
+      'From the four great staves planted in the foreground there is a great garland suspended; two female figures uplift nosegays; at their side is a bridge over a moat, leading to an old manorial house.',
+    numeral: '4',
+  },
+  54: {
+    name: 'Five of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Calmly expressing a dissenting opinion',
+      'Allowing someone to use his or her own methods to get a job done',
+      'Opening the floor for discussion or debate',
+      'Comparing progress made so far to standards set earlier',
+    ],
+    reversedMeanings: [
+      'Berating others for their ridiculous opinions',
+      'Picking fights',
+      'Offering destructive criticism',
+      'Baiting people with barbed remarks',
+      'Disrupting progress with an endless stream of pointless objections',
+    ],
+    suit: 'Wands',
+    keywords: [
+      'confrontation',
+      'disruption',
+      'distinction',
+      'objection',
+      'strife',
+    ],
+    img: 'w05.jpg',
+    symbolism: 'TODO',
+    description:
+      'A posse of youths, who are brandishing staves, as if in sport or strife. It is mimic warfare, and hereto correspond the',
+    numeral: '5',
+  },
+  55: {
+    name: 'Six of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Outperforming your peers',
+      'Winning a competition',
+      'Being recognized as a capable person',
+      'Having your "moment in the spotlight"',
+      'Being cheered on by the crowd',
+      'Getting an award',
+      'Earning the admiration of others',
+      'Telling someone, "Good job!"',
+    ],
+    reversedMeanings: [
+      'Being a bad winner',
+      'Allowing your achievements to inflate your ego',
+      'Looking down on people who seem less capable',
+      'Craving to be the center of attention',
+      'Giving or receiving insincere praise',
+      'Envying the achievements of others',
+    ],
+    suit: 'Wands',
+    keywords: ['victory', 'achievement', 'success', 'triumph'],
+    img: 'w06.jpg',
+    symbolism: 'TODO',
+    description:
+      'A laurelled horseman bears one staff adorned with a laurel crown; footmen with staves are at his side.',
+    numeral: '6',
+  },
+  56: {
+    name: 'Seven of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Refusing to be silenced through fear or intimidation',
+      'Continuing a fight against all odds',
+      'Being fierce',
+      'Defending yourself against physical and emotional attacks',
+      'Refusing to put up with abuse',
+      'Clinging to your values despite all pressure to abandon them',
+    ],
+    reversedMeanings: [
+      'Having a chip on your shoulder',
+      'Taking unnecessary risks as a means of proving your fearlessness',
+      'Looking for an opportunity to take offense',
+      'Responding to constructive criticism with defensiveness',
+      'Refusing to stand up for yourself and your beliefs',
+    ],
+    suit: 'Wands',
+    keywords: ['bravery', 'resolve', 'determination'],
+    img: 'w07.jpg',
+    symbolism: 'TODO',
+    description:
+      'A young man on a craggy eminence brandishing a staff; six other staves are raised towards him from below.',
+    numeral: '7',
+  },
+  57: {
+    name: 'Eight of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Taking swift action',
+      'Moving forward with a plan as quickly as possible',
+      'Energizing yourself',
+      'Adapting to sudden changes',
+      'Taking setbacks in stride',
+      'Embracing the idea that nothing stays the same forever',
+      'Reacting quickly and appropriately to unforeseen problems',
+    ],
+    reversedMeanings: [
+      'Giving in to panic',
+      'Running in circles and screaming',
+      'Insisting things must always stay the same',
+      'Stirring the pot just to see what will happen',
+      'Rushing others',
+      "Refusing to re-evaluate a schedule or program, even when it's clearly no longer appropriate",
+    ],
+    suit: 'Wands',
+    keywords: ['speed', 'swiftness', 'responsiveness', 'change'],
+    img: 'w08.jpg',
+    symbolism: 'TODO',
+    description:
+      'The card represents motion through the immovable-a flight of wands through an open country; but they draw to the term of their course. That which they signify is at hand; it may be even on the threshold.',
+    numeral: '8',
+  },
+  58: {
+    name: 'Nine of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Sticking with it for the duration',
+      'Fulfilling your promises and obligations',
+      'Bearing up under incredible duress',
+      'Dragging yourself across the finish line',
+      'Picking yourself up by your own bootstraps',
+      'Refusing to quit',
+      'Going as far as you can go and being satisfied with your performance',
+    ],
+    reversedMeanings: [
+      'Making yourself a martyr',
+      'Abandoning your post',
+      'Giving up at the first sign of opposition',
+      'Being prevented from fulfilling an obligation',
+      'Failing to be dependable',
+      'Refusing to let something go that needs to be released',
+      'Beating a dead horse',
+    ],
+    suit: 'Wands',
+    keywords: ['toughness', 'persistence', 'stamina', 'loyalty', 'release'],
+    img: 'w09.jpg',
+    symbolism: 'TODO',
+    description:
+      'The figure leans upon his staff and has an expectant look, as if awaiting an enemy. Behind are eight other staves--erect, in orderly disposition, like a palisade.',
+    numeral: '9',
+  },
+  59: {
+    name: 'Ten of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Holding your own in extreme circumstances',
+      'Helping others carry their burdens',
+      'Coming to the aid of the oppressed',
+      'Knowing and being honest about your own limits',
+      'Recognizing when you are not well-suited for a particular task',
+    ],
+    reversedMeanings: [
+      'Taking on more work than you know you can handle',
+      'Refusing to say "No" when you\'re already overloaded',
+      'Making a habit of working overtime',
+      'Shielding others from facing the consequences of their own poor judgment',
+      'Over-extending yourself on a regular basis',
+    ],
+    suit: 'Wands',
+    keywords: ['exhaustion', 'resistance', 'burden', 'oppression'],
+    img: 'w10.jpg',
+    symbolism: 'TODO',
+    description:
+      'A man oppressed by the weight of the ten staves which he is carrying.',
+    numeral: '10',
+  },
+  60: {
+    name: 'Page of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Leaping at a new opportunity',
+      'Being a cheerleader or ardent advocate for your cause',
+      'Being a True Believer',
+      'Taking first steps toward independence',
+      'Trusting in your own abilities',
+      'Asking for feedback',
+    ],
+    reversedMeanings: [
+      'Basing your entire self-image on what others think',
+      'Seizing every new idea that comes your way without question',
+      'Habitually discounting input or feedback from others',
+      'Being so eager to "do it yourself" that you hinder your own progress',
+    ],
+    suit: 'Wands',
+    keywords: [
+      'enthusiasm',
+      'eagerness',
+      'confidence',
+      'validation',
+      'affirmation',
+    ],
+    img: 'w11.jpg',
+    symbolism: 'TODO',
+    description:
+      'In a scene similar to the former, a young man stands in the act of proclamation. He is unknown but faithful, and his tidings are strange.',
+    numeral: '11',
+  },
+  61: {
+    name: 'Knight of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Charging ahead',
+      'Making rapid progress',
+      'Refusing limits',
+      'Dazzling those around you with your wit and charm',
+      'Convincing others of your right to leadership',
+      'Convincing others to follow you',
+      'Being a catalyst for change',
+    ],
+    reversedMeanings: [
+      'Blundering forward with inadequate skill or information',
+      'Running roughshod over the feelings of others',
+      'Using sex appeal to manipulate others',
+      'Forcing your leadership or ideology on others',
+      'Beginning many projects without finishing any',
+    ],
+    suit: 'Wands',
+    keywords: ['boldness', 'bravado', 'passion', 'persuasion', 'advocacy'],
+    img: 'w12.jpg',
+    symbolism: 'TODO',
+    description:
+      'He is shewn as if upon a journey, armed with a short wand, and although mailed is not on a warlike errand. He is passing mounds or pyramids. The motion of the horse is a key to the character of its rider, and suggests the precipitate mood, or things connected therewith.',
+    numeral: '12',
+  },
+  62: {
+    name: 'Queen of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Paying close attention',
+      'Helping others focus on the issue at hand',
+      'Getting everyone to work together',
+      'Identifying common ground',
+      'Bringing people together, despite their differences',
+      'Using reverse psychology',
+    ],
+    reversedMeanings: [
+      'Being distracted, or using your charms or skills to distract others from the goal',
+      'Calling attention to yourself with negative or unhealthy behaviors',
+      'Disrupting group activities as a means of feeding your own ego',
+    ],
+    suit: 'Wands',
+    keywords: ['attention', 'attraction', 'unification', 'collaboration'],
+    img: 'w13.jpg',
+    symbolism: 'TODO',
+    description:
+      "The Wands throughout this suit are always in leaf, as it is a suit of life and animation. Emotionally and otherwise, the Queen's personality corresponds to that of the King, but is more magnetic.",
+    numeral: '13',
+  },
+  63: {
+    name: 'King of Wands',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Putting old things together in new and exciting ways',
+      'Coming up with unexpected solutions',
+      'Using your experience to solve puzzles and problems',
+      'Doing what you set out to do',
+      'Directing the efforts of others',
+    ],
+    reversedMeanings: [
+      'Using your creativity to get out of honest work',
+      'Investing great energy in avoiding responsibility',
+      'Boasting about achievements without putting your expertise to practical use',
+      'Lording it over others',
+    ],
+    suit: 'Wands',
+    keywords: ['creativity', 'ingenuity', 'achievement', 'direction'],
+    img: 'w14.jpg',
+    symbolism: 'TODO',
+    description:
+      'The physical and emotional nature to which this card is attributed is dark, ardent, lithe, animated, impassioned, noble. The King uplifts a flowering wand, and wears, like his three correspondences in the remaining suits, what is called a cap of maintenance beneath his crown. He connects with the symbol of the lion, which is emblazoned on the back of his throne.',
+    numeral: '14',
+  },
+  64: {
+    name: 'Ace of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Outlining a plan for achieving prosperity',
+      'Becoming aware of opportunities to improve income or health',
+      'Realizing you have everything you need',
+      'Appreciating everything the Universe has given you',
+      'Receiving the perfect gift at the perfect time',
+    ],
+    reversedMeanings: [
+      'Indulging in relentless consumerism',
+      'Wanting more, no matter how much you have',
+      'Obsessing on your account balance',
+      'Suffering from hypochondria',
+      'Consuming blessings without expressing gratitude',
+      'Taking what you want without concern for the needs of others',
+    ],
+    suit: 'Pentacles',
+    keywords: ['health', 'wealth', 'practicality', 'receiving'],
+    img: 'p01.jpg',
+    symbolism: 'TODO',
+    description:
+      'A hand--issuing, as usual, from a cloud--holds up a pentacle.',
+    numeral: '1',
+  },
+  65: {
+    name: 'Two of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Weighing options',
+      'Comparing prices',
+      'Determining the value of one option over another',
+      'Juggling resources to make ends meet',
+      "Making difficult choices based on what's best for your body or your bankbook",
+      'Looking at the bottom line',
+      'Asking for a second opinion on health issues',
+    ],
+    reversedMeanings: [
+      'Engaging in endless price comparison',
+      'Putting off a buying decision for fear of finding a slightly better value later on',
+      'Buying something without regard for value',
+      'Breaking your budget with unnecessary expenses',
+      'Engaging in behavior with no regard for how your body or bankbook will be impacted',
+    ],
+    suit: 'Pentacles',
+    keywords: ['evaluation', 'decision', 'budgeting', 'diagnosis'],
+    img: 'p02.jpg',
+    symbolism: 'TODO',
+    description:
+      'A young man, in the act of dancing, has a pentacle in either hand, and they are joined by that endless cord which is like the number 8 reversed.',
+    numeral: '2',
+  },
+  66: {
+    name: 'Three of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Finishing a project',
+      'Setting and meeting standards',
+      'Performing according to specifications',
+      'Making something others value',
+      'Creating something new',
+      'Doing your part in a group project',
+      'Delivering exactly what others have asked for',
+    ],
+    reversedMeanings: [
+      'Pandering to the tastes of others',
+      "Failing to deliver what you've promised",
+      'Not delivering your best work unless closely supervised',
+      'Ignoring or breaking agreements with those who have invested in you',
+      'Refusing to do your part',
+      'Failing to abide by a clearly-outlined agreement with yourself or others',
+    ],
+    suit: 'Pentacles',
+    keywords: ['expression', 'production', 'work', 'contribution'],
+    img: 'p03.jpg',
+    symbolism: 'TODO',
+    description:
+      'A sculptor at his work in a monastery. Compare the design which illustrates the Eight of Pentacles. The apprentice or amateur therein has received his reward and is now at work in earnest.',
+    numeral: '3',
+  },
+  67: {
+    name: 'Four of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Saving for a rainy day',
+      'Fasting as part of a spiritual practice',
+      'Dieting in an effort to improve your body',
+      'Abstaining from sex as a way of honoring a spiritual tradition or personal promise',
+      'Being financially conservative',
+      'Establishing a trust fund',
+      'Opening a savings account',
+    ],
+    reversedMeanings: [
+      'Being stingy',
+      'Refusing to spend money that needs to be spent',
+      'Withholding sex from your partner',
+      'Taking care of your own needs exclusively, without regard for the needs of others',
+      'Spending a dollar to save a penny',
+      "Failing to be a good manager of the blessings you've been given",
+    ],
+    suit: 'Pentacles',
+    keywords: ['protection', 'conservation', 'preservation', 'safety'],
+    img: 'p04.jpg',
+    symbolism: 'TODO',
+    description:
+      'A crowned figure, having a pentacle over his crown, clasps another with hands and arms; two pentacles are under his feet. He holds to that which he has.',
+    numeral: '4',
+  },
+  68: {
+    name: 'Five of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Recognizing your needs and taking action to fulfill them',
+      'Doing as much as you can do with what little you have',
+      'Admitting you need help',
+      'Embracing the aid that comes your way',
+      "Focusing on what you have versus what you don't",
+      'Looking for the light at the end of the tunnel',
+    ],
+    reversedMeanings: [
+      'Exaggerating your financial or physical needs',
+      'Adopting a poverty mentality',
+      'Refusing to support yourself',
+      'Refusing offers of support',
+      'Playing the martyr',
+      'Turning down opportunities to improve your health or finances',
+      'Wallowing in misery',
+    ],
+    suit: 'Pentacles',
+    keywords: ['poverty', 'destitution', 'need', 'crisis'],
+    img: 'p05.jpg',
+    symbolism: 'TODO',
+    description: 'Two mendicants in a snow-storm pass a lighted casement.',
+    numeral: '5',
+  },
+  69: {
+    name: 'Six of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Giving time, money, or effort to a charity',
+      'Taking part in a group effort',
+      'Lending your resources to others without expecting anything in return',
+      'Making sure everyone is treated equally',
+      'Working together toward a common goal',
+      'Redistributing wealth, time, or attention',
+      'Tithing',
+      'Sharing credit for your success',
+    ],
+    reversedMeanings: [
+      'Making a loan as a means of gaining control over someone',
+      'Using charitable acts to draw attention to yourself',
+      'Dividing work or resources unfairly',
+      'Failing to do your part in a group effort',
+      'Ignoring obligations and commitments',
+    ],
+    suit: 'Pentacles',
+    keywords: ['charity', 'fairness', 'cooperation', 'sharing'],
+    img: 'p06.jpg',
+    symbolism: 'TODO',
+    description:
+      'A person in the guise of a merchant weighs money in a pair of scales and distributes it to the needy and distressed. It is a testimony to his own success in life, as well as to his goodness of heart.',
+    numeral: '6',
+  },
+  70: {
+    name: 'Seven of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Measuring progress toward your goal',
+      'Looking at results with an eye toward improving performance',
+      'Asking, "How happy am I?"',
+      'Coming up with ideas for improving your health or prosperity',
+      "Deciding it's time for a change",
+      'Expressing an honest opinion',
+    ],
+    reversedMeanings: [
+      'Becoming distracted by melancholy thoughts',
+      'Longing for "the good old days"',
+      'Beating yourself up over lost opportunities',
+      'Judging your own work harshly',
+      'Holding others to inappropriate standards',
+      'Refusing to take part in a project, then whining about the quality of the outcome',
+    ],
+    suit: 'Pentacles',
+    keywords: ['assessment', 'evaluation', 're-evaluation', 'reflection'],
+    img: 'p07.jpg',
+    symbolism: 'TODO',
+    description:
+      'A young man, leaning on his staff, looks intently at seven pentacles attached to a clump of greenery on his right; one would say that these were his treasures and that his heart was there.',
+    numeral: '7',
+  },
+  71: {
+    name: 'Eight of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Doing your best',
+      'Bringing enthusiasm and zeal to your work',
+      'Making an effort to be the best you can be',
+      'Finding the work that is right for you',
+      'Taking care of the small details',
+      'Becoming a finely skilled craftsperson',
+      'Building something with your hands',
+      'Making a handmade gift',
+    ],
+    reversedMeanings: [
+      'Working yourself to death',
+      'Doing a half-hearted or sloppy job',
+      'Continuing in a job you hate',
+      'Buying thoughtless gifts',
+      'Producing work with shoddy craftsmanship',
+      'Rushing through your work',
+      'Rejecting opportunities to learn more about your craft',
+    ],
+    suit: 'Pentacles',
+    keywords: ['effort', 'work diligence', 'skill'],
+    img: 'p08.jpg',
+    symbolism: 'TODO',
+    description:
+      'An artist in stone at his work, which he exhibits in the form of trophies.',
+    numeral: '8',
+  },
+  72: {
+    name: 'Nine of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Investing time in learning or teaching a difficult task',
+      'Restraining yourself from physical or financial extremes',
+      'Making sacrifices as a way of achieving larger goals',
+      'Breaking a complex task down into simple steps',
+      'Wanting what you have',
+      'Knowing the difference between needs and wants',
+    ],
+    reversedMeanings: [
+      'Being assigned to a task without being trained to perform it',
+      'Pursuing a position for which you are not qualified',
+      'Disregarding requirements',
+      'Refusing to dedicate adequate time or attention when learning about something or someone new',
+      'Always craving more',
+    ],
+    suit: 'Pentacles',
+    keywords: ['training', 'discipline', 'confidence', 'enough'],
+    img: 'p09.jpg',
+    symbolism: 'TODO',
+    description:
+      'A woman, with a bird upon her wrist, stands amidst a great abundance of grapevines in the garden of a manorial house. It is a wide domain, suggesting plenty in all things. Possibly it is her own possession and testifies to material well-being.',
+    numeral: '9',
+  },
+  73: {
+    name: 'Ten of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Celebrating your physical and financial blessings',
+      'Realizing how lucky or how blessed you are',
+      'Being satisfied with your physical and financial achievements',
+      'Taking best advantage of "times of plenty"',
+      'Enjoying a feast',
+      'Showering friends or family with gifts',
+    ],
+    reversedMeanings: [
+      'Spending all of your money on extravagant gifts and possessions',
+      'Trying too hard to impress others with your wealth or physique',
+      'Giving an inappropriately expensive gift as a means of currying favor',
+      'Obsessing on matters of weight, health, or finance',
+      'Always asking, "What\'s in it for me?"',
+    ],
+    suit: 'Pentacles',
+    keywords: ['wealth', 'abundance', 'acquisition', 'greed'],
+    img: 'p10.jpg',
+    symbolism: 'TODO',
+    description:
+      "A man and woman beneath an archway which gives entrance to a house and domain. They are accompanied by a child, who looks curiously at two dogs accosting an ancient personage seated in the foreground. The child's hand is on one of them.",
+    numeral: '10',
+  },
+  74: {
+    name: 'Page of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Learning the value of a dollar',
+      'Starting a savings plan',
+      'Taking the first steps toward getting out of debt',
+      'Learning new physical tasks',
+      'Discovering your sexuality',
+      'Launching a diet, a weight-lifting program, or a health-related effort',
+      'Learning by doing',
+    ],
+    reversedMeanings: [
+      'Trying to appear healthier or wealthier than you really are',
+      'Spending money carelessly',
+      'Living strictly for today, with no thought of tomorrow',
+      'Possessing immature attitudes toward sex and sexuality',
+      'Using wealth or beauty as an excuse for not having to learn and grow',
+    ],
+    suit: 'Pentacles',
+    keywords: [
+      'practicality',
+      'prosperity',
+      'learning',
+      'growth',
+      'adolescence',
+    ],
+    img: 'p11.jpg',
+    symbolism: 'TODO',
+    description:
+      'A youthful figure, looking intently at the pentacle which hovers over his raised hands. He moves slowly, insensible of that which is about him.',
+    numeral: '11',
+  },
+  75: {
+    name: 'Knight of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Spending money wisely',
+      'Saving for a rainy day',
+      'Paying close attention to physical or financial details',
+      'Knowing where every dollar goes',
+      'Having safe sex',
+      'Preferring facts to "good feelings"',
+      'Finding creative ways to "make do" with resources on hand',
+      'Completing a new invention',
+    ],
+    reversedMeanings: [
+      'Throwing caution to the four winds',
+      'Spending without regard for consequence',
+      'Spending on luxury when necessities are lacking',
+      'Escaping stress by spending money',
+      'Obsessing on tiny physical or financial details',
+      'Perpetually chasing after some new bauble',
+      "Copying another's work and claiming it as your own",
+    ],
+    suit: 'Pentacles',
+    keywords: ['caution', 'focus', 'realism', 'invention'],
+    img: 'p12.jpg',
+    symbolism: 'TODO',
+    description:
+      'He rides a slow, enduring, heavy horse, to which his own aspect corresponds. He exhibits his symbol, but does not look therein.',
+    numeral: '12',
+  },
+  76: {
+    name: 'Queen of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Appreciating fine food, fine wine, beautiful art, beautiful bodies, or any of the better things in life',
+      'Reveling in healthy sexuality',
+      'Treating yourself',
+      'Splurging on the occasional "nice to have" item',
+      'Rewarding someone with compensation above and beyond expectations',
+      'Having it all',
+    ],
+    reversedMeanings: [
+      'Indulging in gluttony or greediness',
+      'Becoming insatiable',
+      'Blunting the impact of treats by indulging in them too often',
+      'Providing physical comfort without providing for emotional needs',
+      "Allowing a feeling of entitlement to distort your gratitude for what you're given",
+    ],
+    suit: 'Pentacles',
+    keywords: [
+      'luxury',
+      'comfort',
+      'resourcefulness',
+      'generosity',
+      'prosperity',
+    ],
+    img: 'p13.jpg',
+    symbolism: 'TODO',
+    description:
+      'The face suggests that of a dark woman, whose qualities might be summed up in the idea of greatness of soul; she has also the serious cast of intelligence; she contemplates her symbol and may see worlds therein.',
+    numeral: '13',
+  },
+  77: {
+    name: 'King of Pentacles',
+    arcana: 'Minor Arcana',
+    uprightMeanings: [
+      'Becoming debt-free',
+      'Having more than enough to get by',
+      'Making contributions to a savings plan',
+      'Taking a new job with an eye toward advancing your career',
+      'Buying life or health insurance',
+      'Being confident in the bedroom',
+      'Taking on the role of enforcer when called upon to do so',
+    ],
+    reversedMeanings: [
+      'Becoming so conservative you resist all change on principle alone',
+      'Ignoring innovations in the name of preserving tradition',
+      'Being smug or cocky',
+      'Becoming ruthlessly dedicated to profit or pleasure',
+      'Being sexually selfish',
+      "Bossing others around, especially when you're not empowered to do so",
+    ],
+    suit: 'Pentacles',
+    keywords: ['stability', 'dependability', 'confidence', 'intervention'],
+    img: 'p14.jpg',
+    symbolism: 'TODO',
+    description:
+      "The figure calls for no special description the face is rather dark, suggesting also courage, but somewhat lethargic in tendency. The bull's head should be noted as a recurrent symbol on the throne. The sign of this suit is represented throughout as engraved or blazoned with the pentagram, typifying the correspondence of the four elements in human nature and that by which they may be governed. In many old Tarot packs this suit stood for current coin, money, deniers. I have not invented the substitution of pentacles and I have no special cause to sustain in respect of the alternative. But the consensus of divinatory meanings is on the side of some change, because the cards do not happen to deal especially with questions of money.",
+    numeral: '14',
+  },
+};
+
+// async () => {
+//     const response = await fetch('./output.json');
+//     cardDictionary = await response.json();
+// }
+
+module.exports = cardDictionary;
